Fix `accession_document_association` loading
Before this change, the code would skip adding rows with a document_id
already in that table in the DB. But, all SEC filings have certain
shared docs, and the point here is to show all associations between
filings and the docs they reference. So new rows should only be skipped
when there's already a row in the DB with that (accession_id,
document_id) pair.